Separate YouTube UI and Player. New components and sequencing
